backend/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiofiles==23.2.1

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Async Tasks
celery==5.3.4
redis==5.0.1
pytz==2023.3

# Validation & Settings
pydantic[email]==2.5.0
pydantic-settings==2.1.0
email-validator>=2.0.0

# Security & Authentication
python-jose[cryptography]==3.3.0

# PDF Generation
reportlab==4.0.7
qrcode[pil]==7.4.2
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# OAuth
httpx==0.25.2

# Environment
python-dotenv==1.0.0
urllib3>=2.6.0


# File Processing (New)
python-multipart==0.0.6

# Real-time & Multimedia
python-socketio==5.11.0
boto3==1.34.0
